+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -13,6 +13,16 @@ exclude: true
 
 ⚠️  breaking change and deprecation notices
 
+## v87
+- ⚠ Added [`WebExtension.DownloadInitData`][87.1] class that can be used to
+  implement the WebExtension `downloads` API. This class represents initial state of a download.
+- Added [`WebExtension.Download.Info`][87.2] interface that can be used to
+  implement the WebExtension `downloads` API. This interface allows communicating
+  download's state to Web Extension.
+  
+[87.1]: {{javadoc_uri}}/WebExtension.DownloadInitData.html
+[87.2]: {{javadoc_uri}}/WebExtension.Download.Info.html
+
 ## v86
 - Removed deprecated [`ContentDelegate#onExternalResponse(GeckoSession, WebResponseInfo)`].
   Use [`ContentDelegate#onExternalResponse(GeckoSession, WebResponse)`][82.2] instead.
@@ -877,4 +887,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 437ce82f72ccd40f18d7b7e6f5c0f7e1f6645c02
+[api-version]: d751d69857ce777d1cab63a431bd6d67dc8b2a48